+++ mopidy/mixers/dummy.py
@@ -8,8 +8,8 @@ class DummyMixer(ThreadingActor, BaseMixer):
     def __init__(self):
         self._volume = None
 
-    def _get_volume(self):
+    def get_volume(self):
         return self._volume
 
-    def _set_volume(self, volume):
+    def set_volume(self, volume):
         self._volume = volume