+++ tests/frontends/mpd/dispatcher_test.py
@@ -1,33 +1,33 @@
 import unittest
 
 from mopidy.backends.dummy import DummyBackend
-from mopidy.frontends.mpd import dispatcher
+from mopidy.frontends.mpd.dispatcher import MpdDispatcher
 from mopidy.frontends.mpd.exceptions import MpdAckError
-from mopidy.frontends.mpd.protocol import request_handlers, handle_pattern
+from mopidy.frontends.mpd.protocol import request_handlers, handle_request
 from mopidy.mixers.dummy import DummyMixer
 
 class MpdDispatcherTest(unittest.TestCase):
     def setUp(self):
-        self.b = DummyBackend.start().proxy()
+        self.backend = DummyBackend.start().proxy()
         self.mixer = DummyMixer.start().proxy()
-        self.h = dispatcher.MpdDispatcher()
+        self.dispatcher = MpdDispatcher()
 
     def tearDown(self):
-        self.b.stop().get()
+        self.backend.stop().get()
         self.mixer.stop().get()
 
     def test_register_same_pattern_twice_fails(self):
         func = lambda: None
         try:
-            handle_pattern('a pattern')(func)
-            handle_pattern('a pattern')(func)
+            handle_request('a pattern')(func)
+            handle_request('a pattern')(func)
             self.fail('Registering a pattern twice shoulde raise ValueError')
         except ValueError:
             pass
 
     def test_finding_handler_for_unknown_command_raises_exception(self):
         try:
-            self.h.find_handler('an_unknown_command with args')
+            self.dispatcher._find_handler('an_unknown_command with args')
             self.fail('Should raise exception')
         except MpdAckError as e:
             self.assertEqual(e.get_mpd_ack(),
@@ -37,18 +37,18 @@ class MpdDispatcherTest(unittest.TestCase):
         expected_handler = lambda x: None
         request_handlers['known_command (?P<arg1>.+)'] = \
             expected_handler
-        (handler, kwargs) = self.h.find_handler('known_command an_arg')
+        (handler, kwargs) = self.dispatcher._find_handler('known_command an_arg')
         self.assertEqual(handler, expected_handler)
         self.assert_('arg1' in kwargs)
         self.assertEqual(kwargs['arg1'], 'an_arg')
 
     def test_handling_unknown_request_yields_error(self):
-        result = self.h.handle_request('an unhandled request')
+        result = self.dispatcher.handle_request('an unhandled request')
         self.assertEqual(result[0], u'ACK [5@0] {} unknown command "an"')
 
     def test_handling_known_request(self):
         expected = 'magic'
         request_handlers['known request'] = lambda x: expected
-        result = self.h.handle_request('known request')
+        result = self.dispatcher.handle_request('known request')
         self.assert_(u'OK' in result)
         self.assert_(expected in result)