+++ tests/frontends/mpd/exception_test.py
@@ -1,7 +1,7 @@
 import unittest
 
-from mopidy.frontends.mpd.exceptions import (MpdAckError, MpdUnknownCommand,
-    MpdSystemError, MpdNotImplemented)
+from mopidy.frontends.mpd.exceptions import (MpdAckError, MpdPermissionError,
+    MpdUnknownCommand, MpdSystemError, MpdNotImplemented)
 
 class MpdExceptionsTest(unittest.TestCase):
     def test_key_error_wrapped_in_mpd_ack_error(self):
@@ -25,10 +25,9 @@ class MpdExceptionsTest(unittest.TestCase):
 
     def test_get_mpd_ack_with_values(self):
         try:
-            raise MpdAckError('A description', error_code=6, index=7,
-                command='foo')
+            raise MpdAckError('A description', index=7, command='foo')
         except MpdAckError as e:
-            self.assertEqual(e.get_mpd_ack(), u'ACK [6@7] {foo} A description')
+            self.assertEqual(e.get_mpd_ack(), u'ACK [0@7] {foo} A description')
 
     def test_mpd_unknown_command(self):
         try:
@@ -43,3 +42,10 @@ class MpdExceptionsTest(unittest.TestCase):
         except MpdSystemError as e:
             self.assertEqual(e.get_mpd_ack(),
                 u'ACK [52@0] {} foo')
+
+    def test_mpd_permission_error(self):
+        try:
+            raise MpdPermissionError(command='foo')
+        except MpdPermissionError as e:
+            self.assertEqual(e.get_mpd_ack(),
+                u'ACK [4@0] {foo} you don\'t have permission for "foo"')